refactor: move draggablelist to generic folder
